starting the reconfig
