+++ src/routes.js
@@ -1,0 +1,17 @@
+// import React from "react";
+// import { Route, IndexRoute } from "react-router";
+// import Home from "./home/index";
+
+// // Components
+// import App from "./components/app";
+// import { Guide, Docs } from "./screens/docs/index";
+// import About from "./screens/about/index";
+
+// module.exports = (
+//   <Route path="/" component={App}>
+//     <IndexRoute component={Home}/>
+//     <Route path="/docs" component={Docs}/>
+//     <Route path="/docs/getting-started" component={Guide}/>
+//     <Route path="/about" component={About}/>
+//   </Route>
+// );
